pyproject.toml
[tool.poetry]
name = "extrator-leads"
version = "0.1.0"
description = "Extrator e enriquecedor de leads com interface web moderna"
authors = ["Seu Nome <seu.email@exemplo.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
pandas = "^1.3.0"
googlemaps = "^4.4.0"
requests = "^2.26.0"
beautifulsoup4 = "^4.9.3"
python-dotenv = "^0.19.0"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.7b0"
isort = "^5.9.3"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 